+++ services/ui-src/src/components/MeasureWrapper/index.tsx
@@ -69,6 +69,12 @@ export const MeasureWrapper = ({ measure, name, year, measureId }: Props) => {
       >
         <form data-testid="measure-wrapper-form">
           <CUI.Container maxW="4xl" as="section">
+            <CUI.Text fontSize="sm">
+              For technical questions regarding use of this application, please
+              reach out to MDCT_Help@cms.hhs.gov. For content-related questions
+              about measure specifications, or what information to enter in each
+              field, please reach out to MACQualityTA@cms.hhs.gov.
+            </CUI.Text>
             {cloneElement(measure, {
               name,
               year,
