+++ services/ui-src/src/views/CoreSet/index.tsx
@@ -51,14 +51,15 @@ export const CoreSet = () => {
       ],
     },
     {
-      Type: "Health Home",
-      title: "Admission to an Institution from the Community",
-      abbr: "AIF-HH",
-      path: `/${state}/${year}/${coreSetId}/AIF-HH`,
+      Type: "Adult",
+      title:
+        "Follow-Up After Emergency Department Visit for Alcohol and Other Drug Abuse or Dependence",
+      abbr: "FUA-AD",
+      path: `/${state}/${year}/${coreSetId}/FUA-AD`,
       isReporting: false,
       rateComplete: 0,
       lastDateModified: "",
-      id: "AIF-HH",
+      id: "FUA-AD",
       actions: [
         {
           itemText: "Edit",
